Projeto finalizado!

+++ src/main/java/com/aulasJava/DesafioCrudClientes/services/ClientService.java
@@ -1,15 +1,19 @@
 package com.aulasJava.DesafioCrudClientes.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.aulasJava.DesafioCrudClientes.ResourceNotFoundException;
 import com.aulasJava.DesafioCrudClientes.DTO.ClientDTO;
+import com.aulasJava.DesafioCrudClientes.Exceptions.DatabaseException;
+import com.aulasJava.DesafioCrudClientes.Exceptions.ResourceNotFoundException;
 import com.aulasJava.DesafioCrudClientes.entities.Client;
 import com.aulasJava.DesafioCrudClientes.repositories.ClientRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -37,14 +41,26 @@ public class ClientService {
 
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO clientDto) {
-		Client client = repository.getReferenceById(id);
-		DtoToClient(clientDto, client);
-		return new ClientDTO(repository.save(client));
+		try {
+			Client client = repository.getReferenceById(id);
+			DtoToClient(clientDto, client);
+			return new ClientDTO(repository.save(client));
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+
 	}
 
 	@Transactional
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha na integridade referencial!");
+		}
 	}
 
 	public void DtoToClient(ClientDTO dto, Client client) {
